Adding starting point for Kent moodle theme
Using Aardvark as a base, this new theme is the starting point for the new Kent theme. The point of doing this now is for me to learn a little about how to do themes so that I can better esitmate how long it will take to complete. Also this will give the flts a better version of Moodle to demo new things about Moodle 2 with